requirements.txt
langchain>=1.2.13
langfuse>=3.0,<4.0
langchain-openai>=1.1.12
langchain-community>=0.4.1
python-dotenv>=1.2.2
langgraph-checkpoint-mongodb>=0.3.1
fastapi>=0.135
uvicorn>=0.42
httpx>=0.28
PyJWT[crypto]>=2.12
pymongo>=4.0
azure-search-documents>=11.4
azure-core>=1.30
openai>=1.0
